+++ packages/core/src/workflow/internal/utils.ts
@@ -10,7 +10,6 @@ import type { WorkflowStreamController } from "../stream";
 import type { WorkflowStreamEvent } from "../types";
 import type { WorkflowState } from "./state";
 import type {
-  InternalExtractWorkflowInputData,
   InternalWorkflowStateParam,
   InternalWorkflowStepConfig,
   WorkflowExecuteContext,
@@ -69,11 +68,12 @@ export function defaultStepConfig<CONFIG extends InternalWorkflowStepConfig>(con
  * @returns The execution context for the step
  */
 export function createStepExecutionContext<INPUT, DATA, SUSPEND_DATA, RESUME_DATA>(
-  data: InternalExtractWorkflowInputData<DATA>,
+  data: DATA,
   state: InternalWorkflowStateParam<INPUT>,
   executionContext: WorkflowExecutionContext,
   suspendFn: (reason?: string, suspendData?: SUSPEND_DATA) => Promise<never>,
   resumeData?: RESUME_DATA,
+  retryCount = 0,
 ): WorkflowExecuteContext<INPUT, DATA, SUSPEND_DATA, RESUME_DATA> {
   return {
     data,
@@ -81,6 +81,7 @@ export function createStepExecutionContext<INPUT, DATA, SUSPEND_DATA, RESUME_DAT
     getStepData: (stepId: string) => executionContext?.stepData.get(stepId),
     suspend: suspendFn,
     resumeData,
+    retryCount,
     logger: executionContext.logger,
     writer: executionContext.streamWriter,
   };